requirements.txt
gradio
python-dotenv
pandas
pypdf
loguru
promptic
tenacity
PyPDF2
# PyPDF2==3.0.1

# AI stuff
openai

# RENDER.io stuff
fastapi
uvicorn
celery
redis 

# AWS Stuff
boto3
botocore

# Supadase Stuff
supabase 
